Move query logic to service and queries modules
