+++ app/filters/common.py
@@ -1,5 +1,6 @@
 import uuid
 from datetime import datetime
+from typing import Annotated
 
 from fastapi_filter import FilterDepends, with_prefix
 
@@ -67,12 +68,23 @@ class AgentFilter(CustomFilter):
         ordering_model_fields = ["pref_label"]  # noqa: RUF012
 
 
-class ContributionFilterMixin:
-    contribution: AgentFilter | None = FilterDepends(with_prefix("contribution", AgentFilter))
-
-
 class CreationFilterMixin:
     creation_date__lte: datetime | None = None
     creation_date__gte: datetime | None = None
     update_date__lte: datetime | None = None
     update_date__gte: datetime | None = None
+
+
+# Dependencies
+MTypeClassFilterDep = Annotated[MTypeClassFilter, FilterDepends(MTypeClassFilter)]
+ETypeClassFilterDep = Annotated[ETypeClassFilter, FilterDepends(ETypeClassFilter)]
+SpeciesFilterDep = Annotated[SpeciesFilter, FilterDepends(SpeciesFilter)]
+StrainFilterDep = Annotated[StrainFilter, FilterDepends(StrainFilter)]
+AgentFilterDep = Annotated[AgentFilter, FilterDepends(AgentFilter)]
+
+# Nested dependencies
+NestedMTypeClassFilterDep = FilterDepends(with_prefix("mtype", MTypeClassFilter))
+NestedETypeClassFilterDep = FilterDepends(with_prefix("etype", ETypeClassFilter))
+NestedSpeciesFilterDep = FilterDepends(with_prefix("species", SpeciesFilter))
+NestedStrainFilterDep = FilterDepends(with_prefix("strain", StrainFilter))
+NestedAgentFilterDep = FilterDepends(with_prefix("contribution", AgentFilter))
