+++ app/filters/morphology.py
@@ -1,28 +1,46 @@
-from datetime import datetime
+from typing import Annotated
 
 from fastapi_filter import FilterDepends, with_prefix
 
 from app.db.model import ReconstructionMorphology
 from app.filters.base import CustomFilter
-from app.filters.common import AgentFilter, MTypeClassFilter, SpeciesFilter, StrainFilter
+from app.filters.common import (
+    AgentFilter,
+    CreationFilterMixin,
+    MTypeClassFilter,
+    NestedAgentFilterDep,
+    NestedMTypeClassFilterDep,
+    NestedSpeciesFilterDep,
+    NestedStrainFilterDep,
+    SpeciesFilter,
+    StrainFilter,
+)
 
 
-class MorphologyFilter(CustomFilter):
-    creation_date__lte: datetime | None = None
-    creation_date__gte: datetime | None = None
-    update_date__lte: datetime | None = None
-    update_date__gte: datetime | None = None
+class MorphologyFilter(
+    CustomFilter,
+    CreationFilterMixin,
+):
     name__ilike: str | None = None
     brain_region_id: int | None = None
     species_id__in: list[int] | None = None
 
-    mtype: MTypeClassFilter | None = FilterDepends(with_prefix("mtype", MTypeClassFilter))
-    species: SpeciesFilter | None = FilterDepends(with_prefix("species", SpeciesFilter))
-    strain: StrainFilter | None = FilterDepends(with_prefix("strain", StrainFilter))
-    contribution: AgentFilter | None = FilterDepends(with_prefix("contribution", AgentFilter))
+    mtype: Annotated[MTypeClassFilter | None, NestedMTypeClassFilterDep] = None
+    species: Annotated[SpeciesFilter | None, NestedSpeciesFilterDep] = None
+    strain: Annotated[StrainFilter | None, NestedStrainFilterDep] = None
+    contribution: Annotated[AgentFilter | None, NestedAgentFilterDep] = None
 
     order_by: list[str] = ["-creation_date"]  # noqa: RUF012
 
     class Constants(CustomFilter.Constants):
         model = ReconstructionMorphology
         ordering_model_fields = ["creation_date", "update_date", "name"]  # noqa: RUF012
+
+
+# Dependencies
+MorphologyFilterDep = Annotated[MorphologyFilter, FilterDepends(MorphologyFilter)]
+# Nested dependencies
+NestedMorphologyFilterDep = FilterDepends(with_prefix("morphology", MorphologyFilter))
+NestedExemplarMorphologyFilterDep = FilterDepends(
+    with_prefix("exemplar_morphology", MorphologyFilter)
+)
