+++ app/service/organization.py
@@ -1,0 +1,50 @@
+import uuid
+
+import sqlalchemy as sa
+
+from app.db.model import Organization
+from app.dependencies.common import PaginationQuery
+from app.dependencies.db import SessionDep
+from app.errors import ensure_result
+from app.schemas.agent import OrganizationCreate, OrganizationRead
+from app.schemas.types import ListResponse, PaginationResponse
+
+
+def read_many(
+    db: SessionDep,
+    pagination_request: PaginationQuery,
+) -> ListResponse[OrganizationRead]:
+    query = sa.select(Organization)
+
+    data = db.execute(
+        query.offset(pagination_request.offset).limit(pagination_request.page_size)
+    ).scalars()
+
+    total_items = db.execute(query.with_only_columns(sa.func.count(Organization.id))).scalar_one()
+
+    response = ListResponse[OrganizationRead](
+        data=[OrganizationRead.model_validate(d) for d in data],
+        pagination=PaginationResponse(
+            page=pagination_request.page,
+            page_size=pagination_request.page_size,
+            total_items=total_items,
+        ),
+        facets=None,
+    )
+
+    return response
+
+
+def read_one(id_: uuid.UUID, db: SessionDep) -> OrganizationRead:
+    with ensure_result(error_message="Organization not found"):
+        stmt = sa.select(Organization).filter(Organization.id == id_)
+        row = db.execute(stmt).scalar_one()
+    return OrganizationRead.model_validate(row)
+
+
+def create(organization: OrganizationCreate, db: SessionDep) -> OrganizationRead:
+    row = Organization(**organization.model_dump())
+    db.add(row)
+    db.commit()
+    db.refresh(row)
+    return OrganizationRead.model_validate(row)
